Gateway now only accepts path packets from first-tier nodes
Upon receiving a path-returning packet, gateway first checks if this
packet is from a first-tier node. The packet is discarded if the sender
is not at the first tier.

This guards against cross-tier snooping, which may cause a lower-tier
node to have no upper-tier node when the path packet from the upper-tier
node is lost, but the one from the lower-tier one to the upper-tier one
is accidentally snooped by the gateway.

The above error will let the LabVIEW server crash.

Some unecessary signals and slots that explicitly state first-tier node
receiving events are deleted, as ALL path packet events are now
implicitly oriented from first-tier nodes.